# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank REST API
  description: REST API для управления банковскими картами, пользователями и транзакциями
  version: 4.7.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер
security:
  - bearerAuth: [ ]
paths:
...
  /api/admin/cards/{id}:
    get:
      summary: Получение карты по ID (админ)
      description: Получение информации о карте по её ID
      operationId: getCardById
      tags:
        - Карты (Админ)
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о карте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Обновление карты (админ)
      description: Обновление данных карты администратором
      operationId: updateCard
      tags:
        - Карты (Админ)
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardUpdateRequest'
      responses:
        '200':
          description: Карта обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Удаление карты (админ)
      description: Удаление карты по ID
      operationId: deleteCard
      tags:
        - Карты (Админ)
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Карта удалена
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/user/cards:
    get:
      summary: Получение карт пользователя
      description: Получение списка карт текущего пользователя
      operationId: getUserCards
      tags:
        - Карты (Пользователь)
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardResponse'
...
components:
  schemas:
...
    CardUpdateRequest:
      type: object
      properties:
        cardNumber:
          type: string
          description: Номер карты
        expiryDate:
          type: string
          description: Срок действия карты (YYYY-MM)
        cvv:
          type: string
          description: CVV код
        balance:
          type: number
          format: double
          description: Баланс карты
      required:
        - cardNumber
        - expiryDate
        - cvv
        - balance
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
        userId:
          type: integer
          format: int64
          description: ID пользователя
        cardNumber:
          type: string
          description: Номер карты
        expiryDate:
          type: string
          description: Срок действия карты (YYYY-MM)
        balance:
          type: number
          format: double
          description: Баланс карты
      required:
        - id
        - userId
        - cardNumber
        - expiryDate
        - balance
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Сообщение об ошибке
        status:
          type: integer
          description: HTTP статус код
      required:
        - message
        - status
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT